# openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "OpsPilot AI Control Tower API",
    "description": "Autonomous IT Operations Control Tower powered by AI, OpenTelemetry, and intelligent incident management.",
    "version": "0.1.0"
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Enter your Bearer JWT token obtained from POST /api/v1/auth/login"
      }
    },
...
  },
  "paths": {
...
    "/api/v1/simulator/": {
      "post": {
        "summary": "Inject chaos failure scenario into target service",
        "tags": [
          "simulator"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "serviceId",
                  "scenario"
                ],
                "properties": {
                  "serviceId": {
                    "type": "string"
                  },
                  "scenario": {
                    "type": "string",
                    "enum": [
                      "BAD_DEPLOYMENT",
                      "HIGH_CPU",
                      "MEMORY_LEAK",
                      "DB_CONNECTION_EXHAUSTION",
                      "API_LATENCY",
                      "QUEUE_BACKLOG",
                      "BATCH_FAILURE",
                      "DISK_FULL",
                      "DEPENDENCY_FAILURE",
                      "CERT_EXPIRY"
                    ]
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      }
    },
    "/api/v1/simulator/chaos": {
      "post": {
        "summary": "Inject chaos failure scenario into target service",
        "tags": [
          "simulator"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "serviceId",
                  "scenario"
                ],
                "properties": {
                  "serviceId": {
                    "type": "string"
                  },
                  "scenario": {
                    "type": "string",
                    "enum": [
                      "BAD_DEPLOYMENT",
                      "HIGH_CPU",
                      "MEMORY_LEAK",
                      "DB_CONNECTION_EXHAUSTION",
                      "API_LATENCY",
                      "QUEUE_BACKLOG",
                      "BATCH_FAILURE",
                      "DISK_FULL",
                      "DEPENDENCY_FAILURE",
                      "CERT_EXPIRY"
                    ]
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      }
    },
    "/api/v1/simulator/heal": {
      "post": {
        "summary": "Heal injected chaos failure scenarios",
        "tags": [
          "simulator"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "serviceId": {
                    "type": "string",
                    "description": "Optional target service ID to heal, or omit to heal all"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      }
    },
...
  },
  "servers": [
    {
      "url": "http://localhost:3001",
      "description": "Local Development Server"
    },
    {
      "url": "https://opspilotapi-production.up.railway.app",
      "description": "Production Server"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    }
  ]
}